pyproject.toml
[tool.poetry]
name = "neta"
version = "0.3.0"
description = "WhatsApp-AI chat integration bridge"
authors = ["Leonardo Fukui <leofukui@hotmail.com>"]
readme = "README.md"
packages = [{include = "neta", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.29.0"
webdriver-manager = "^4.0.0"
pyperclip = "^1.9.0"
python-dotenv = "^1.0.0"
requests = "^2.32.3"
pillow = "^11.2.1"
openai = "^1.12.0"
anthropic = "^0.17.0"
google-generativeai = "^0.3.0"
httpx = "0.27.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.3"
pytest = "^7.4.3"
black = "^23.11.0"

[tool.poetry.scripts]
neta = "neta.main:main"
format = "black src:main src"
lint = "ruff check src:main"
test = "pytest:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
select = ["E", "F", "W", "I", "N", "B"]
ignore = ["E501"]
target-version = "py310"
src = ["src"]

[tool.black]
line-length = 120
target-version = ["py310"]
include = '\.pyi?$'
